Revert "buffer/file name completion parity"
While this is appealing for consistency, I don't actually want it. Until I find myself
working on a linux box in an emacs session with many similarly named but different cased
files open in multiple buffers on a regular basis, the convenience in 99.9% of the use
cases outweighs the benefits of conceptual consistency.