disable gc when in minibuffer
Packages that are memory-intensive in the minibuffer (e.g. fuzzyfinders like flx) have
found they can get a performance boost by increasing the garbage-collection threshold. The
problem with setting a high gc threshold and leaving it high is that a lot of garbage
collects over time and the eventual cleanup causes a noticeable lag. We can tune
performance by disabling gc during such memory-intensive tasks, then resetting it to the
low default value of 800kb.