fixing problems with hydra-scroll
To prevent the cursor from jumping halfway up the page when scrolling
takes it past the bottom of the window, make hydra use the echo area for
hints instead of constructing a new window. This works as long as the
hint isn't long enough to force line-wrapping. Mostly hide the hints anyway.