drag-stuff-except-modes org-mode
The Org documentation makes it sounds like the org bindings for
reordering/promoting/demoting should be available when the cursor is on
a heading, and the global bindings (in this case drag-stuff) should be
available when the cursor is on a body line. This doesn't seem to be the
case though -- maybe drag-stuff is doing something that too forcefully
overrides those keys. Until I can figure out a way to get org bindings
on headings and drag-stuff on body lines, I'm just disabling drag-stuff
in org buffers.