crazy wgrep workaround
Messing with the whitespace in the ag results buffer breaks wgrep. Set
things up so that whitespace is trimmed by default, and when wgrep is
engaged the results are first recompiled with the raw whitespace
included.

This implementation relies on adding and removing advice and hooks. I
think it could be made more straightforward by remapping
wgrep-change-to-wgrep-mode to a wrapper function.