toggle-compilation-window: basic ag
This is the major use-case I wanted to support -- much like the quickfix
list in Vim. It should really be broken out into its own package. Next
step is handling modes like rspec-mode, which I keep in a separate frame
if on a multi-monitor setup, but want to toggle like the quickfix if I'm
using a single frame.