fix ivan/rotate-theme
The modeline colors were getting mangled because I was enabling the
new theme before disabling the old ones. Instead of enabling the new
theme and then doing extra work to exclude it from the disabling step,
make a backup from which to restore the original themes if enabling
the new one fails, and attempt to enable new theme after disabling the
current ones.