fix doom-buffer-mode/text-scale-mode conflict
doom-buffer-mode makes 'face-remapping-alist buffer-local in order to
customize faces on a per-buffer basis. When enabled, it marks the
variable as permanent-local so that its customizations will persist
when normal (non-permanent) local variables are wiped out (e.g. when
the major-mode changes).

Similarly, text-scale-mode uses 'face-remapping-alist as a local
variable to scale the height of the buffer's text. It also uses it's
own 'text-scale-mode-remapping local variable to remember what scaling
it has added to 'face-remapping-alist. This lets it reset the text to
its default height by finding and removing its customizations from the
alist.

Because text-scale-mode assumes 'face-remapping-alist is *not*
permanent-local, it allows 'text-scale-mode-remapping to be wiped out
whenever non-permanent local variables are cleared. The assumption is
that clearing local variables will clear 'face-remapping-alist as
well, removing any scaling that text-scale-mode had applied, so
there's no need to preserve the value of 'text-scale-mode-remapping.

Because doom-buffer-mode makes 'face-remapping-alist permanent-local,
whenever local variables are cleared out, we lose track of what text
scaling was applied, but the scaling itself remains in effect. It's
still stored in 'face-remapping-alist and we no longer have to cookie
with which to identify and remove it. Any scaling currently in effect
becomes the new baseline; we can continue to increase/decrease the
scaling manually, but attempting to reset it to its default just
resets it to this new baseline.

I ran into this when scaling a help buffer and then quitting its
window. The next time I brought that help buffer up, it had the
previous scaling in effect and behaved as if that was the default
height.