magnet-names
Add support for buffer-name matching, so we can give magnetic
treatment to buffers like "*Pp Eval Output*".

In its current form, this treats the elements of magnet-names as exact
strings to match, not regular expressions. So far, the majority of
use-cases are covered by checking the major-mode, and any cases that
require resorting to inspecting the buffer-name are likely to contain
asterisk characters, which would force the user to litter their
magnet-names setq expressions with backslashes were we to treat them
as regular expressions. If we find we want to add support for regular
expressions, we'll have to chose between providing a separate
variable (e.g. magnet-patterns), squeezing a plist into the mix like
shackle.el does (i.e. `:regexp t`), or forcing users to add the
necessary backslashes to their setq expressions.