magnet: wrestle pp-eval-output to the ground
It would be nice if magnet.el provided the user with a way to specify
such "sibling" features/behavior, as in, "Any buffer that qualifies
for magnetic treatment should also have this other feature applied to
it, and here's when you should check". On the other hand, maybe that's
coupling things at the wrong layer.