hydra-dired-follow
Implement the equivalent of follow-mode for previewing files from dired. The
hoops I had to jump through to add a remap to the transient keymap are not worth
the effort for this use-case (I have to explicitly bind C-n and C-p anyway
because I've bound them to some dual-commands), but I wanted to record it this
way in the commit for reference, in case I find a use-case that really does call
out for applying a remap to a hydra-generated keymap.