fix evil-state in view-mode
The way evil tries to define an initial state for view-mode doesn't work,
because view-mode is a minor mode and evil's initial-state mechanism works on
major modes.

I only discovered this because I wanted to bind "q" in view-mode, and using
evil-define-key to do so yeilded strange results: upon first attempt (whether
typing "q" or asking emacs to describe that key), it was bound to
evil-record-macro. After hitting escape or q, it would then change to be bound
to my custom command.

Using motion state makes more sense in a view-mode context, and it circumvents
the "q" keybinding conflict. Use view-mode-hook to set the evil-state appropriately.