dired-preview-clean: uncouple window deletion from buffer killing
To understand the need for this change, consider the case where user has file ~/A
open in a single window. They switch that window to dired in the parent
directory (~/), and preview ~/B. This splits their window and displays a preview
of ~/B in the new split. The user then moves the cursor onto the dired entry for
~/A and triggers preview. The preview window switches from ~/B (in view-mode) to
~/A (not in view-mode). Then the user triggers dired-preview-clean: buffer A
will NOT be killed, since it's not in view-mode, but the preview window SHOULD
still be deleted, since it was created explicitly for dired-preview.