put installed packages under version-control
There are some updates available for evil, doom-theme, and smartparens. These are pretty
core to my config, so I'd like to be able to rollback if something doesn't work out in the
new versions. I've also been meaning to put the entire packages directory under version
control for this reason for a while. Now's probably a good time to do it!