ruby-mode: explicitly set underscore as word char
It's not enough to rely on change-major-mode-hook in the case of ruby-mode,
because ruby-mode explicitly sets underscore to be a symbol char.

Another solution would be to move my setting to after-change-major-mode-hook,
since that runs after ruby-mode has already done its work. I think for now,
though, I'd rather allow specific modes to override my default, and override the
overrides as I come across them and decide to. If this keeps happening and I
still can't imagine a case where it would be useful to have underscore as
anything other than a word char, I'll reconsider using after-change-major-mode-hook.